fix(settings): use testNotificationProvider server action with robust error handling and fallback validation
